models/unet++/requirements.txt
# Core DL Framework
torch>=2.0.0
torchvision>=0.15.0

# Model Architecture
segmentation-models-pytorch>=0.3.3

# Image Processing & Augmentation
albumentations>=1.3.1
numpy>=1.24.0
Pillow>=9.5.0
opencv-python-headless>=4.8.0

# Utilities
tqdm>=4.65.0
wandb>=0.15.0